+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
-* May 3, 2014
+* May 9, 2014
+
+  - Updated ``example/Blackjack.py``.
 
   - Updated ``example/Memory.py``.